ace_it/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==2.2
django-celery==3.3.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-tastypie==0.14.2
importlib-metadata==0.23
kombu==3.0.37
more-itertools==7.2.0
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0

#for chat/ messaging
aioredis==1.2.0
asgiref==2.3.2
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
certifi==2018.10.15
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
constantly==15.1.0
daphne==2.2.2
defusedxml==0.5.0
django-allauth==0.38.0
django-cors-headers==2.4.0
gunicorn==19.7.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.5.6
oauthlib==2.1.0
psycopg2-binary==2.7.5
PyHamcrest==1.9.0
pylint==2.1.1
python3-openid==3.1.0
requests==2.20.1
requests-oauthlib==1.0.0
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.0
urllib3==1.25.2
whitenoise==3.3.1
wrapt==1.10.11
zope.interface==4.6.0
